+++ src/Collections/Transactions.php
@@ -4,7 +4,6 @@ namespace GingerPluginSdk\Collections;
 
 use GingerPluginSdk\Entities\Transaction;
 use GingerPluginSdk\Interfaces\AbstractCollectionContainerInterface;
-use JetBrains\PhpStorm\Pure;
 
 final class Transactions extends AbstractCollection implements AbstractCollectionContainerInterface
 {
@@ -16,9 +15,13 @@ final class Transactions extends AbstractCollection implements AbstractCollectio
         foreach ($items as $item) {
             $this->add($item);
         }
-        parent::__construct($item, 'transactions');
+        parent::__construct( 'transactions');
     }
 
+    /**
+     * @param \GingerPluginSdk\Entities\Transaction $transaction
+     * @return $this
+     */
     public function addTransaction(Transaction $transaction): Transactions
     {
         $this->add($transaction);